requirements.txt
Flask==2.0.2
firebase-admin==5.2.0
langchain==0.1.0
langchain-community==0.2.0
requests==2.25.1
gunicorn==20.1.0  # Serveur WSGI pour Heroku
python-dotenv==0.19.0  # Pour charger des variables d'environnement depuis un fichier .env
